+++ src/components/UsageBar/index.tsx
@@ -3,6 +3,7 @@ import * as PropTypes from 'prop-types';
 import { ProgressBar } from 'react-bootstrap';
 import { bulk, xbytes } from '../../utils';
 import { UsageBarProps } from '../../interface';
+import lang from "../../locale/language";
 import './style.scss';
 
 const PERCENT_WITH_STATUS = {
@@ -39,6 +40,7 @@ const UsageBar: React.FC<UsageBarProps> = props => {
     isExcludeWarning,
     waterLine,
     isHideFooter,
+    formatType,
   } = props;
   const hasNow = props.hasOwnProperty('now');
   const hasPercent = props.hasOwnProperty('percent');
@@ -47,8 +49,8 @@ const UsageBar: React.FC<UsageBarProps> = props => {
   const errorPercent = props.unavailableData && max && props.unavailableData / max;
   let nowValue: number | string | undefined = now;
   let maxValue: number | string | undefined = max;
-  let nowSuffix = '';
-  let maxSuffix = '';
+  let nowSuffix: any = '';
+  let maxSuffix: any = '';
   let left;
   let right;
   let bsStyle;
@@ -60,16 +62,21 @@ const UsageBar: React.FC<UsageBarProps> = props => {
     nowSuffix = maxSuffix = '%';
   }
 
-  if (isByte) {
+  if (isByte || formatType) {
     // hasPercent 表明左边数据设置完成（百分比形式），不需要再调整
+    let byteOptions: object = { splitUnit: true };
+    if (formatType) {
+      byteOptions = { splitUnit: true, formatType };
+    }
+
     if (!hasPercent && now) {
-      const nowArr: any = xbytes(now, { splitUnit: true });
+      const nowArr: any = xbytes(now, byteOptions);
       nowValue = nowArr[0];
       nowSuffix = nowArr[1];
     }
-    const maxArr: any = max && xbytes(max, { splitUnit: true });
-    maxValue = maxArr[0];
-    maxSuffix = maxArr[1];
+    const maxArr: any = max && xbytes(max, byteOptions);
+    maxValue = maxArr[0] === undefined ? 0 : maxArr[0];
+    maxSuffix = maxArr[1] === undefined ? 'B' : maxArr[1];
   } else if (isBulk) {
     // hasPercent 表明左边数据设置完成（百分比形式），不需要再调整
     if (!hasPercent && now) {
@@ -78,15 +85,15 @@ const UsageBar: React.FC<UsageBarProps> = props => {
       nowSuffix = nowArr[1] as string;
     }
     const maxArr: any = max && bulk(max, { splitUnit: true });
-    maxValue = maxArr[0];
-    maxSuffix = maxArr[1];
+    maxValue = maxArr[0] === undefined ? 0 : maxArr[0];
+    maxSuffix = maxArr[1] === undefined ? '' : maxArr[1];
   }
 
   // max 为 0 时，设置为无限制
   // isPercent，hasPercent 涉及到百分比展示，此时 max 不应该为无限制
   // showZeroMax 直接展示 max 为处理后的 0 + 单位
   if (!max && !hasPercent && !isPercent && !showZeroMax) {
-    maxValue = '无限制';
+    maxValue = lang().MAX_VALUE;
     maxSuffix = '';
   }
 
@@ -163,6 +170,10 @@ UsageBar.propTypes = {
    * 数字以字节（B, KB, MB, GB...）为单位展示
    **/
   isByte: PropTypes.bool,
+  /**
+   * 格式化输出，可选[decimal|binary]
+   **/
+  formatType: PropTypes.string,
   /**
    * 数字以数量（万, 亿, 兆, 京...）为单位展示
    **/