+++ src/locale/en.ts
@@ -1,0 +1,4 @@
+export default {
+    MORE_TITLE: 'More',
+    MAX_VALUE: 'no limit',
+}