+++ src/locale/zh.ts
@@ -1,0 +1,4 @@
+export default {
+    MORE_TITLE: '更多',
+    MAX_VALUE: '无限制',
+}